refactor(proxy): hoist liveKeysAtPath + isArrayPath to proxy-live-keys (CORE-P1c)
The two helpers were byte-identical between `errors-proxy.ts` and
`field-state-proxy.ts` (~32 LOC duplicated; both surfaces drive their
container `ownKeys` / array-shape decision off the live form value).
Move them into `core/proxy-live-keys.ts` as the single source so the
two callsites import from one place.

No observable behavior change: same `getAtPath(state.form.value, …)`
read inside the consumer's active effect; same numeric-string keys for
array values; same `[]` for primitives + nullish; same `false` for
`isArrayPath` at the root segment. Verified by the existing 119
proxy-surface tests across `values-proxy`, `surface-proxy`,
`field-state-proxy`, `wizard-statuses-proxy`, `proxy-write-traps`,
`errors-proxy-enumeration`, `array-proxy-methods`.
