+++ src/runtime/core/field-state-proxy.ts
@@ -9,6 +9,7 @@ import {
 } from './field-state-api'
 import { getAtPath } from './path-walker'
 import type { Path, Segment } from './paths'
+import { isArrayPath, liveKeysAtPath } from './proxy-live-keys'
 import { buildSurfaceProxy, type SurfaceProxy } from './surface-proxy'
 
 /**
@@ -203,47 +204,6 @@ export function buildFieldStateProxy<F extends GenericForm>(
     containerOwnKeys: (segments) => liveKeysAtPath(state, segments),
     isArrayContainer: (segments) => isArrayPath(state, segments),
   })
-}
-
-/**
- * Live keys for the form data at a container path. Powers
- * `Object.keys(form.fields.items)` and `v-for` iteration over array
- * field proxies. The read happens inside the consumer's active
- * effect so `state.form.value` is tracked: appending or removing
- * items re-enumerates on the next render.
- */
-function liveKeysAtPath<F extends GenericForm>(
-  state: FormStore<F, GenericForm>,
-  segments: readonly Segment[]
-): readonly string[] {
-  const value = getAtPath(state.form.value, segments)
-  if (value === null || value === undefined) return []
-  if (Array.isArray(value)) {
-    const keys = new Array<string>(value.length)
-    for (let i = 0; i < value.length; i += 1) keys[i] = String(i)
-    return keys
-  }
-  if (typeof value === 'object') return Object.keys(value as Record<string, unknown>)
-  return []
-}
-
-/**
- * Whether the path resolves to an array container RIGHT NOW. The
- * live form value is the source of truth here — a discriminated
- * union variant switch can swap the path between array-shaped and
- * something else, and the proxy's reported shape needs to track
- * what the consumer can actually see, not what the schema's loosest
- * variant could permit. The container cache keys off this same
- * predicate (see `containerProxyAt` in surface-proxy.ts), so a
- * shape flip surfaces a freshly-targeted proxy on the next read
- * through `form.fields.X`.
- */
-function isArrayPath<F extends GenericForm>(
-  state: FormStore<F, GenericForm>,
-  segments: readonly Segment[]
-): boolean {
-  if (segments.length === 0) return false
-  return Array.isArray(getAtPath(state.form.value, segments))
 }
 
 /**
